Refactor chart utils into separate utils directory; util file naming nits for tooltip/axis label converters
