+++ src/client/utils/charts/computeChartDimensionDomain.ts
