+++ src/client/utils/charts/computeScatterLineEndpoints.ts
