+++ src/client/utils/charts/dataPointToTooltipLabel.ts
